requirements.txt
fastapi==0.112.0
uvicorn==0.30.5
sqlalchemy==2.0.32
psycopg2-binary==2.9.9
pytest==8.3.2
alembic==1.13.2
python-decouple==3.8
